+++ src/Exceptions/InvalidKeyException.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SnowmanNunu\Weather\Exceptions;
+
+class InvalidKeyException extends HttpException
+{
+}